requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.99.9
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2